LIST: [PATCH 4/10] rust: xarray: add `XArrayState`
Add `XArrayState` as internal state for XArray iteration and entry
operations. This struct wraps the C `xa_state` structure and holds a
reference to a `Guard` to ensure exclusive access to the XArray for the
lifetime of the state object.

The `XAS_RESTART` constant is also exposed through the bindings helper
to properly initialize the `xa_node` field.

The struct and its constructor are marked with `#[expect(dead_code)]` as
there are no users yet. We will remove this annotation in a later patch.

Link: https://lore.kernel.org/r/20251203-xarray-entry-send-v1-4-9e5ffd5e3cf0@kernel.org
Signed-off-by: Andreas Hindborg <a.hindborg@kernel.org>